% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{add.gen}
\alias{add.gen}
\title{Add gen key to inp}
\usage{
add.gen(inp)
}
\arguments{
\item{inp}{Input list}

\item{ny}{Number of grid cells in y-direction}

\item{Dx}{Diffusivity in x-direction}

\item{Dy}{Diffusivity in y-direction}

\item{land}{Vector containing linear indices of land cells.}
}
\value{
Generator as sparse matrix.
}

