% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalikelihood.R
\name{calc.data.likelihood}
\alias{calc.data.likelihood}
\title{Calculate data likelihood}
\usage{
calc.data.likelihood(inp)
}
\arguments{
\item{parname}{Character string containing the name of the variable of interest.}

\item{rep}{A result report as generated by running fit.shmm.}

\item{exp}{Take exp of the variable? TRUE/FALSE.}
}
\value{
Input list containing combined data likelihood of all data sources. Format is a matrix where nrow is equal to number of time steps, and ncol is equal to the total number of grid cells.
}
\details{
Calculate data likelihood
}
\examples{
inp <- calc.data.likelihood(inp)
}

