% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/checkinp.R
\name{check.inp}
\alias{check.inp}
\title{Check list of input variables}
\usage{
check.inp(inp)
}
\arguments{
\item{inp}{List of input variables, see details for required variables.}
}
\value{
An updated list of input variables checked for consistency and with defaults added.
}
\description{
Check list of input variables
}
\details{
Check list of input variables
}
\examples{
inp <- check.inp(inp)
}

