% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generator.R
\name{diag.sparse}
\alias{diag.sparse}
\title{Create sparse identity matrix}
\usage{
diag.sparse(n)
}
\arguments{
\item{n}{Dimension of matrix}
}
\value{
Sparse identity matrix
}
\description{
Create sparse identity matrix
}

