% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{find.land}
\alias{find.land}
\title{Function to remove lakes etc.}
\usage{
find.land(nas, do.plot = FALSE)
}
\arguments{
\item{nas}{Matrix containing NAs at land cells}

\item{do.plot}{Plot progress}
}
\value{
Matrix containing TRUE at land and FALSE at connected water.
}

