% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shmm.R
\name{fit.shmm}
\alias{fit.shmm}
\title{Fit a continuous-time surplus production model to data.}
\usage{
fit.shmm(inp, dbg = 0)
}
\arguments{
\item{inp}{List of input variables as output by check.inp.}

\item{dbg}{Debugging option. Will print out runtime information useful for debugging if set to 1. Will print even more if set to 2.}
}
\value{
List containing results.
}
\description{
Fit a continuous-time surplus production model to data.
}
\details{
Fit a continuous-time surplus production model to data.
}
\examples{
rep <- fit.shmm(inp)
}

