% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{format.distr}
\alias{format.distr}
\title{Format distribution from flattened time x vector (2D) to time x matrix (3D).}
\usage{
\method{format}{distr}(vecdistr, rep)
}
\arguments{
\item{vecdistr}{Distribution to be formatted}

\item{rep}{Result of running fit.shmm.}
}
\value{
The requested distribution as an 3D-array.
}

