% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.distr}
\alias{get.distr}
\title{Extract distribution from rep}
\usage{
get.distr(name, rep)
}
\arguments{
\item{name}{Name of the distribution to plot. Either 'phi', 'pred' or 'smoo'.}

\item{rep}{Result of running fit.shmm.}
}
\value{
The requested distribution as an 3D-array.
}

