% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.mean.track}
\alias{get.mean.track}
\title{Calculate an estimated track from the mean of the smoothed distribution.}
\usage{
get.mean.track(rep)
}
\arguments{
\item{rep}{Result of running fit.shmm.}
}
\value{
A list containing
}

