% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.par}
\alias{get.par}
\title{Extract parameters from a result report as generated by fit.shmm.}
\usage{
get.par(parname, rep = rep, exp = FALSE)
}
\arguments{
\item{parname}{Character string containing the name of the variable of interest.}

\item{rep}{A result report as generated by running fit.shmm.}

\item{exp}{Take exp of the variable? TRUE/FALSE.}
}
\value{
A matrix with four columns containing respectively: 1) the lower 95% confidence limit; 2) the parameter estimate; 3) the upper 95% confidence limit; 4) the parameter standard deviation in the domain it was estimated (log or non-log).
}
\details{
Helper function for extracting the value and uncertainty of a specific model parameter or derived quantity.
}
\examples{
logD <- get.par('logD', rep, exp=TRUE)
}

