% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{make.generator}
\alias{make.generator}
\title{Create generator}
\usage{
make.generator(nx, ny, Dx, Dy, land = NULL)
}
\arguments{
\item{nx}{Number of grid cells in x-direction}

\item{ny}{Number of grid cells in y-direction}

\item{Dx}{Diffusivity in x-direction}

\item{Dy}{Diffusivity in y-direction}

\item{land}{Vector containing linear indices of land cells.}
}
\value{
Generator as sparse matrix.
}

