% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{make.ns}
\alias{make.ns}
\title{Make skeleton of generator matrix for north-south movement}
\usage{
make.ns(n, nx, land)
}
\arguments{
\item{n}{Total number of grid cells (nx*ny)}

\item{nx}{Number of grid cells in x-direction}

\item{land}{Vector containing linear indices of land cells.}
}
\value{
Skeleton of generator matrix for north-south movement as sparse matrix
}

