% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{max.rate}
\alias{max.rate}
\title{Calculate maximum rate}
\usage{
\method{max}{rate}(Dx, Dy)
}
\arguments{
\item{Dx}{Diffusivity in x-direction}

\item{Dy}{Diffusivity in y-direction}
}
\value{
Maximum rate
}

