% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotshmm.distr}
\alias{plotshmm.distr}
\title{Plots the distribution indicated by name}
\usage{
plotshmm.distr(rep, name = "smoo", sleep = 0.05, add.map = TRUE)
}
\arguments{
\item{rep}{Result of running fit.shmm.}

\item{name}{Name of the distribution to plot. Either 'phi', 'pred' or 'smoo'.}

\item{sleep}{Number of seconds to pause between each frame of the animation.}

\item{add.map}{Add map?}
}
\value{
Noting.
}
\details{
Distribution is plotted as an "animation"
}

