% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{remove.no.access}
\alias{remove.no.access}
\title{Remove cells from generator that are inaccessible}
\usage{
remove.no.access(S, n, nx, land)
}
\arguments{
\item{S}{Generator matrix}

\item{n}{Total number of grid cells (nx*ny)}

\item{nx}{Number of grid cells in x-direction}

\item{land}{Vector containing linear indices of land cells.}
}
\value{
Generator as sparse matrix with inaccessible cells removed.
}

