% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{set.default}
\alias{set.default}
\title{Set default parameter value}
\usage{
set.default(inpin, key, val)
}
\arguments{
\item{inpin}{Input list}

\item{key}{Key of list}

\item{val}{Value}
}
\value{
Modified inpin
}

