% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Description.R
\docType{package}
\name{shmm}
\alias{shmm}
\alias{shmm-package}
\title{Fits a spatial hidden Markov model to tag data}
\description{
Fits a spatial hidden Markov model to tag data
}
\examples{
rep <- test.shmm()
}
\author{
Martin W. Pedersen \email{mawp@dtu.dk}
}
\references{
\url{https://github.com/mawp/shmm/}
}
\seealso{
\code{\link{test.shmm}}
}
\keyword{data,}
\keyword{geolocation}
\keyword{movement}

