% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.shmm}
\alias{sim.shmm}
\title{Simulate movement data}
\usage{
sim.shmm(inp, nobs = 40)
}
\arguments{
\item{inp}{List with parameters specified in the ini key and time step in the dt key}

\item{nobs}{Optional specification of the number of simulated observations.}
}
\value{
A list containing the simulated data.
}
\details{
Simulates data
}
\examples{
sim <- sim.shmm(inp)
}

