% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{sumshmm.parest}
\alias{sumshmm.parest}
\title{Parameter estimates of a fit.shmm() run.}
\usage{
sumshmm.parest(rep, numdigits = 8)
}
\arguments{
\item{rep}{A result report as generated by running fit.shmm.}

\item{numdigits}{Present values with this number of digits after the dot.}
}
\value{
data.frame containing parameter estimates.
}
\description{
Parameter estimates of a fit.shmm() run.
}

