% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{test.shmm}
\alias{test.shmm}
\title{Example of a shmm analysis.}
\usage{
test.shmm(nobs = 40)
}
\arguments{
\item{nobs}{Number of observations to simulate}
}
\value{
A result report as given by fit.shmm().
}
\description{
Example of a shmm analysis.
}
\details{
Loads a data set, fits the model, plots the results.
}
\examples{
rep <- test.shmm()
}

