% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generator.R
\name{uniformization}
\alias{uniformization}
\title{Calculate the matrix exponential of a generator matrix using the uniformization algorithm}
\usage{
uniformization(A, dt)
}
\arguments{
\item{A}{Generator matrix}

\item{dt}{Time step}
}
\value{
Transition probability matrix as sparse matrix
}
\description{
Calculate the matrix exponential of a generator matrix using the uniformization algorithm
}

