% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalikelihood.R
\name{calc.data.likelihood}
\alias{calc.data.likelihood}
\title{Calculate data likelihood}
\usage{
calc.data.likelihood(inp)
}
\arguments{
\item{parname}{Character string containing the name of the variable of interest.}

\item{rep}{A result report as generated by running fit.shmm.}

\item{exp}{Take exp of the variable? TRUE/FALSE.}
}
\value{
A matrix with four columns containing respectively: 1) the lower 95% confidence limit; 2) the parameter estimate; 3) the upper 95% confidence limit; 4) the parameter standard deviation in the domain it was estimated (log or non-log).
}
\details{
Calculate data likelihood
}
\examples{
inp <- calc.data.likelihood(inp)
}

