% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shmm.R
\name{make.datlist}
\alias{make.datlist}
\title{Create data list used as input to TMB::MakeADFun.}
\usage{
make.datlist(inp)
}
\arguments{
\item{inp}{List of input variables as output by check.inp.}
}
\value{
List to be used as data input to TMB::MakeADFun.
}
\description{
Create data list used as input to TMB::MakeADFun.
}

