% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shmm.R
\name{make.parlist}
\alias{make.parlist}
\title{Create parameter list used as input to TMB::MakeADFun.}
\usage{
make.parlist(inp)
}
\arguments{
\item{inp}{List of input variables as output by check.inp.}
}
\value{
List to be used as parameter input to TMB::MakeADFun.
}
\description{
Create parameter list used as input to TMB::MakeADFun.
}

