% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.shmmcls}
\alias{summary.shmmcls}
\title{Output a summary of a fit.shmm() run.}
\usage{
\method{summary}{shmmcls}(object, ...)
}
\arguments{
\item{object}{A result report as generated by running fit.shmm.}

\item{...}{additional arguments affecting the summary produced.}
}
\value{
Nothing. Prints a summary to the screen.
}
\description{
Output a summary of a fit.shmm() run.
}
\details{
TBA
}
\examples{
rep <- fit.shmm(inp)
summary(rep)
}

